update time interval changed to float, alpha blending enabled

+++ src/win/platform.cpp
@@ -557,8 +557,8 @@ int APIENTRY WinMain(
 
             // update game state (and animations)
             game.Update(
-                double(currenttime.QuadPart - lasttime.QuadPart) /
-                double(frequency.QuadPart)
+                float(double(currenttime.QuadPart - lasttime.QuadPart) /
+                      double(frequency.QuadPart))
             );
             lasttime = currenttime;
